+++ src/app/Componentes/Usuario/lista-mascotas/lista-mascotas.component.css
@@ -1,0 +1,22 @@
+.table {
+   width: 100%;
+   border-collapse: collapse;
+ }
+ 
+ .table th, .table td {
+   padding: 8px;
+   border: 1px solid #ddd;
+ }
+ 
+ .table th {
+   background-color: #f2f2f2;
+ }
+ 
+ .table tbody tr:nth-child(even) {
+   background-color: #f9f9f9;
+ }
+ 
+ .table tbody tr:hover {
+   background-color: #e9e9e9;
+ }
+
