+++ src/app/Componentes/Usuario/lista-mascotas/lista-mascotas.component.html
@@ -1,0 +1,26 @@
+<table class="table">
+   <thead>
+     <tr>
+       <th>ID Mascota</th>
+       <th>Nombre</th>
+       <th>Raza</th>
+       <th>Edad</th>
+       <th>Género</th>
+       <th>Fecha de Nacimiento</th>
+       <th>Peso</th>
+       <th>Owner ID</th>
+     </tr>
+   </thead>
+   <tbody>
+     <tr *ngFor="let mascota of mascotas">
+       <td>{{ mascota.id_mascota }}</td>
+       <td>{{ mascota.nombre }}</td>
+       <td>{{ mascota.raza }}</td>
+       <td>{{ mascota.edad }}</td>
+       <td>{{ mascota.genero }}</td>
+       <td>{{ mascota.fecha_nacimiento }}</td>
+       <td>{{ mascota.peso }}</td>
+       <td>{{ mascota.Owner_id_owner }}</td>
+     </tr>
+   </tbody>
+ </table>
