--- src/app/Usuario/lista-mascotas/lista-mascotas.component.html
@@ -1,1 +1,0 @@
-<p>lista-mascotas works!</p>
