--- src/app/Usuario/lista-mascotas/lista-mascotas.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-lista-mascotas',
-  templateUrl: './lista-mascotas.component.html',
-  styleUrls: ['./lista-mascotas.component.css']
-})
-export class ListaMascotasComponent {
-
-}
